add custom, overblown component for lua-controlled players
My intuition (which might well be wrong) says that the aspect of being
controlled by execution of Lua code doesn't fit in well with the ECS
architecture: The events that players should receive are especially
hard to "generate" or "observe" when everything is so decoupled via
components, for instance. It would mean lots of reverse lookups of the
actual entities to find everything that might involve players.

So for now I've decided to put everything player-related into the
`lua-player' class. This means that there will be more systems, namely
ones that handle stuff involving players, like collision, movement,
etc., and (sometimes very similar) ones that don't.